# openapi.yaml
openapi: '3.1.0'
info:
    title: Manna AI Agent API
    version: '1.0.0'
    contact:
        name: Manna API
        url: https://github.com/Guebbit/manna
    description: |
        Local-first agentic AI platform. Exposes the agent reasoning loop, IDE-specific
        direct-LLM endpoints, file-upload helpers,
        and a multi-library PDF ingestion and semantic search system.

        **Base URL (default):** `http://localhost:3001`
servers:
    - url: http://localhost:3001
      description: Local development server
...
paths:
...
    /run:
        post:
            tags: [core]
            summary: Submit a task to the agent reasoning loop
            description: |
                Triggers the full agentic loop: model routing → tool selection → execution → repeat
                (up to `MAX_STEPS`, default 5). Returns the agent's final answer as a string.

                This is the right endpoint whenever no specialised endpoint covers the use case.
            operationId: postRun
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RunRequest'
                        examples:
                            basic:
                                summary: Simple read task
                                value:
                                    task: 'What npm scripts are available in this project?'
                            withProfile:
                                summary: Force reasoning profile
                                value:
                                    task: 'Explain the trade-offs of the current memory strategy'
                                    profile: reasoning
                            withWrite:
                                summary: Enable write tools
                                value:
                                    task: "Create a new file at data/hello.txt with the content 'hello'"
                                    allowWrite: true
            responses:
                '200':
                    description: Agent completed successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RunResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /run/stream:
        post:
            tags: [core]
            summary: Submit a task and receive Server-Sent Events
            description: |
                Same as `POST /run` but streams agent lifecycle events in real time as
                Server-Sent Events (SSE).

                The connection stays open until the agent completes (on `done`, `error`, or
                `max_steps` event). The original `POST /run` endpoint is completely unchanged.

                **SSE event types**:
                - `step` — emitted after each LLM response: `{ step, action, thought }`
                - `tool` — emitted after each tool execution: `{ tool, result? }` or `{ tool, error }`
                - `route` — emitted when a model profile is selected: `{ profile, model, reason }`
                - `done` — final answer ready: `{ result }`
                - `error` — agent threw: `{ error }`
                - `max_steps` — loop exhausted: `{ task, summary }`
            operationId: postRunStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RunRequest'
                        examples:
                            basic:
                                summary: Stream a simple read task
                                value:
                                    task: 'List all TypeScript files in packages/agent/'
            responses:
                '200':
                    description: SSE stream — connection kept open until agent completes
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: Newline-delimited SSE frames
                            example: |
                                event: route
                                data: {"profile":"code","model":"qwen2.5-coder:14b-instruct-q8_0","reason":"keyword_match:code"}

                                event: step
                                data: {"step":0,"action":"shell","thought":"I need to list TypeScript files..."}

                                event: tool
                                data: {"tool":"shell","result":"packages/agent/agent.ts\n..."}

                                event: done
                                data: {"result":"The TypeScript files in packages/agent/ are: ..."}

                '400':
                    $ref: '#/components/responses/BadRequest'

    # ── Swarm endpoints ───────────────────────────────────────────────────────
    /run/swarm:
        post:
            tags: [swarm]
            summary: Run a multi-agent swarm to solve a complex task
            description: |
                Decomposes the task into subtasks using an LLM, executes each subtask
                through a specialised agent (respecting dependency ordering), and
                synthesises a final answer from all subtask results.

                Use this instead of `POST /run` when the task is complex enough to
                benefit from divide-and-conquer (e.g. "build X, write tests, and document it").
            operationId: postRunSwarm
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SwarmRequest'
                        examples:
                            basic:
                                summary: Complex multi-step task
                                value:
                                    task: 'Build a React login page with form validation, write unit tests, and create API documentation'
                            withOptions:
                                summary: With profile override and subtask limit
                                value:
                                    task: 'Refactor the auth module and update the related tests'
                                    profile: code
                                    maxSubtasks: 3
                                    allowWrite: true
            responses:
                '200':
                    description: Swarm completed successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SwarmResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /run/swarm/stream:
        post:
            tags: [swarm]
            summary: Run a multi-agent swarm with SSE streaming
            description: |
                Same as `POST /run/swarm`, but streams agent lifecycle events as
                Server-Sent Events. The connection stays open until the swarm completes.

                SSE event types: `decomposed`, `subtask_start`, `subtask_done`,
                `subtask_error`, `step`, `tool`, `route`, `done`, `error`.
            operationId: postRunSwarmStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SwarmRequest'
            responses:
                '200':
                    description: SSE event stream
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    Server-Sent Events stream. Each event is typed
                                    (decomposed, subtask_start, subtask_done, step, tool, route, done, error).
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── Workflow endpoints ─────────────────────────────────────────────────────
    /workflow:
        post:
            tags: [workflow]
            summary: Run an explicit ordered list of steps sequentially
            description: |
                Accepts an array of step task strings and runs each as an independent bounded
                `agent.run()` sub-call.  Each step has its own `maxStepsPerStep` iteration cap
                (defaulting to `AGENTS_MAX_STEPS`), so a misbehaving step cannot consume the
                entire budget.

                ### Context carry modes (`carry`)

                | Mode      | Behaviour |
                |-----------|-----------|
                | `none`    | Steps are fully isolated — no prior context forwarded. |
                | `summary` | A compact bullet-point summary of prior step results is prepended to each subsequent step prompt. **(default)** |
                | `full`    | The complete verbatim output of every prior step is appended. Context may grow quickly. |

                Use `POST /workflow` when your outer loop is already structured as an ordered list
                and you want each step bounded independently.
            operationId: postWorkflow
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WorkflowRequest'
                        examples:
                            basic:
                                summary: Two-step workflow (default carry=summary)
                                value:
                                    steps:
                                        - task: 'List all TypeScript files under packages/'
                                        - task: 'Summarise the purpose of each listed file'
                            withOptions:
                                summary: Multi-step write workflow with per-step cap
                                value:
                                    steps:
                                        - task: 'Read the existing index.ts and summarise its structure'
                                        - task: 'Create a new helper module at src/utils/helper.ts'
                                        - task: 'Write unit tests for the helper module'
                                    allowWrite: true
                                    profile: code
                                    carry: full
                                    maxStepsPerStep: 10
            responses:
                '200':
                    description: 'All steps completed (some may have `success: false` if individual steps errored)'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/WorkflowResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /workflow/stream:
        post:
            tags: [workflow]
            summary: Run a sequential workflow with SSE streaming
            description: |
                Same as `POST /workflow`, but streams lifecycle events as Server-Sent Events.
                The connection stays open until all steps complete.

                SSE event types (in order):
                - `workflow_start` — before the first step; `{ stepCount }`.
                - `step_start`     — before each step; `{ index, task }`.
                - `step`           — inner agent iteration; `{ workflowIndex, step, action, thought }`.
                - `tool`           — tool result/error; `{ workflowIndex, tool, result? | error? }`.
                - `route`          — model routing decision; `{ workflowIndex, profile, model, reason }`.
                - `step_done`      — after each step; full `WorkflowStepResult` object.
                - `done`           — after all steps; full `WorkflowResponse` object.
                - `error`          — on fatal error; `{ error }`.
            operationId: postWorkflowStream
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WorkflowRequest'
            responses:
                '200':
                    description: SSE event stream
                    content:
                        text/event-stream:
                            schema:
                                type: string
                                description: |
                                    Server-Sent Events stream. Each event is typed
                                    (workflow_start, step_start, step, tool, route, step_done, done, error).
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── IDE endpoints ─────────────────────────────────────────────────────────
    /autocomplete:
        post:
            tags: [ide]
            summary: Cursor-time code completion
            description: |
                Single LLM call (no agent loop). Returns one or more code completion suggestions
                for the given cursor position. Uses `TOOL_IDE_MODEL` (default `starcoder2`).

                Intended for IDE integration. Responses are cached for identical prefix/suffix pairs.
            operationId: postAutocomplete
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AutocompleteRequest'
            responses:
                '200':
                    description: Completion result
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AutocompleteResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /lint-conventions:
        post:
            tags: [ide]
            summary: Deterministic lint + LLM convention findings
            description: |
                Runs deterministic TypeScript/style checks and optionally enriches findings
                with LLM explanations. Returns a structured list of findings with severity,
                rule name, line numbers, and suggested fixes.
            operationId: postLintConventions
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LintConventionsRequest'
            responses:
                '200':
                    description: Lint findings
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LintResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'
    /page-review:
        post:
            tags: [ide]
            summary: Whole-file categorised engineering review
            description: |
                Single LLM call that analyses a full source file and returns categorised
                review suggestions (correctness, maintainability, performance, security).
            operationId: postPageReview
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PageReviewRequest'
            responses:
                '200':
                    description: Review result
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PageReviewResponse'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '500':
                    $ref: '#/components/responses/InternalError'

    # ── Upload endpoints ───────────────────────────────────────────────────────
...
    /info/models:
        get:
            tags: [info]
            summary: List models currently available in the connected Ollama instance
            description: |
                Proxies Ollama's `GET /api/tags` endpoint and returns the list of
                locally available models with size, digest, and detail metadata.
                Returns 502 if Ollama is unreachable.
            operationId: getInfoModels
            responses:
                '200':
                    description: Model list from Ollama
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    count:
                                        type: integer
                                        description: Number of models available.
                                    ollamaBaseUrl:
                                        type: string
                                        description: The Ollama base URL queried.
                                        example: 'http://localhost:11434'
                                    models:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                name:
                                                    type: string
                                                    example: 'llama3.1:8b-instruct-q8_0'
                                                size:
                                                    type: integer
                                                    nullable: true
                                                    description: Model file size in bytes.
                                                digest:
                                                    type: string
                                                    nullable: true
                                                modifiedAt:
                                                    type: string
                                                    nullable: true
                                                    description: ISO 8601 timestamp of last modification.
                                                details:
                                                    type: object
                                                    nullable: true
                                                    description: Ollama model metadata (family, parameter_size, quantization_level, etc.).
                '502':
                    description: Ollama unreachable or returned an error.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
...
components:
...
    responses:
        BadRequest:
            description: Invalid request body or parameters
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        InternalError:
            description: Internal server error (LLM failure, Qdrant unavailable, etc.)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        ErrorResponse:
            type: object
            required: [error]
            properties:
                error:
                    type: string
                    description: Human-readable error message.
                    example: "Library 'scientific-american' not found"
...
        RunRequest:
            type: object
            required: [task]
            properties:
                task:
                    type: string
                    description: Natural-language description of what the agent should do.
                    example: 'What npm scripts are available in this project?'
                allowWrite:
                    type: boolean
                    default: false
                    description: |
                        When `true`, unlocks the `write_file`, `scaffold_project`, and `document_ingest` tools.
                        Default `false`.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile, bypassing automatic routing.
                        If omitted, the router selects a profile based on the task text.
        RunResponse:
            type: object
            properties:
                result:
                    type: string
                    description: The agent's final answer as a plain string.

        # ── Swarm schemas ────────────────────────────────────────────────────────
        SwarmRequest:
            type: object
            required: [task]
            properties:
                task:
                    type: string
                    description: Natural-language description of the complex task to solve.
                    example: 'Build a React login page, write tests, and create documentation'
                allowWrite:
                    type: boolean
                    default: false
                    description: When `true`, unlocks write tools for all subtask agents.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile for all subtask agents,
                        overriding the decomposer's per-subtask suggestions.
                maxSubtasks:
                    type: integer
                    minimum: 1
                    maximum: 10
                    default: 6
                    description: Maximum number of subtasks the decomposer may produce.
...
        SwarmResponse:
            type: object
            properties:
                answer:
                    type: string
                    description: The synthesised final answer.
                result:
                    type: string
                    deprecated: true
                    description: Deprecated alias for `answer` kept for backward compatibility.
                subtaskResults:
                    type: array
                    items:
                        $ref: '#/components/schemas/SwarmSubtaskResult'
                    description: Individual results for each subtask.
                decomposition:
                    type: object
                    properties:
                        reasoning:
                            type: string
                            description: Decomposer's rationale for the subtask split.
                        subtaskCount:
                            type: integer
                            description: Number of subtasks in the decomposition.
                totalDurationMs:
                    type: number
                    description: Total wall-clock duration of the entire swarm run.

        # ── Workflow schemas ─────────────────────────────────────────────────────
        WorkflowRequest:
            type: object
            required: [steps]
            properties:
                steps:
                    type: array
                    items:
                        $ref: '#/components/schemas/StepDefinition'
                    minItems: 1
                    maxItems: 50
                    description: |
                        Ordered list of workflow step definitions.
                    example:
                        - task: 'List all TypeScript files under packages/'
                        - task: 'Summarise the purpose of each listed file'
                allowWrite:
                    type: boolean
                    default: false
                    description: |
                        When `true`, unlocks `write_file`, `scaffold_project`, and `document_ingest`
                        tools for all steps.
                profile:
                    type: string
                    enum: [fast, reasoning, code, default]
                    description: |
                        Force a specific model profile for all steps, bypassing automatic routing.
                carry:
                    type: string
                    enum: [none, summary, full]
                    default: summary
                    description: |
                        How prior step outputs are carried into subsequent steps.
                        `none` = isolated; `summary` = compact bullet summary (default);
                        `full` = verbatim full output.
                maxStepsPerStep:
                    type: integer
                    minimum: 1
                    maximum: 100
                    description: |
                        Per-step agent-loop iteration cap.  Overrides `AGENTS_MAX_STEPS` for this
                        request only.  Defaults to `AGENTS_MAX_STEPS` (env var, default 5).
...
        WorkflowResponse:
            type: object
            properties:
                steps:
                    type: array
                    items:
                        $ref: '#/components/schemas/WorkflowStepResult'
                    description: Per-step results in execution order.
                allSucceeded:
                    type: boolean
                    description: '`true` when every step completed successfully.'
                totalDurationMs:
                    type: number
                    description: Total wall-clock duration across all steps in milliseconds.

        # ── IDE schemas ───────────────────────────────────────────────────────────
        AutocompleteRequest:
            type: object
            required: [prefix]
            properties:
                prefix:
                    type: string
                    description: Code text before the cursor.
                suffix:
                    type: string
                    description: Code text after the cursor (FIM context).
                language:
                    type: string
                    description: Programming language identifier (e.g. "typescript", "python").
        AutocompleteResponse:
            type: object
            properties:
                requestId:
                    type: string
                    format: uuid
                completion:
                    type: string
                    description: Suggested code to insert at the cursor.
                cached:
                    type: boolean
                    description: Whether the result was served from cache.
                latencyMs:
                    type: integer
...
        LintConventionsRequest:
            type: object
            required: [code]
            properties:
                code:
                    type: string
                    description: Source code to lint.
                language:
                    type: string
                    description: Programming language identifier.
                filename:
                    type: string
                    description: Optional source file name/path for context.
                includeLlm:
                    type: boolean
                    default: true
                    description: Whether to include the optional LLM enrichment pass.
                model:
                    type: string
                    description: Override the model used for LLM enrichment.
                maxFindings:
                    type: integer
                    minimum: 1
                    maximum: 200
                    default: 80
                    description: Maximum number of findings returned.
...
        PageReviewRequest:
            type: object
            required: [code]
            properties:
                code:
                    type: string
                    description: Full source file content to review.
                language:
                    type: string
                filename:
                    type: string
                projectContext:
                    type: string
                    description: Optional sentence describing the project for context.
        PageReviewResponse:
            type: object
            properties:
                requestId:
                    type: string
                    format: uuid
                findings:
                    type: array
                    items:
                        type: object
                latencyMs:
                    type: integer

        # ── Library schemas ───────────────────────────────────────────────────────
...